Neural_Style_Transfer/requirements.txt
streamlit==1.35.0
transformers==4.41.2
torch
PyPDF2
tensorflow==2.15.0
tensorflow-hub==0.14.0
Pillow==10.4.0
numpy<2.0.0
